+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Version history
 
+## [v0.5.0](https://github.com/agrande/lba2remake/compare/v0.4.0...v0.5.0) Big update!
+_Unreleased_
+
+#### Editor
+
+- Palette: Eye drop colors from any editor's main view.
+
 ## [v0.4.0](https://github.com/agrande/lba2remake/compare/v0.3.0...v0.4.0) Big update!
 _October 10th, 2020_